Using the new managementPermissions flags to determine what parts of the client UI to hide
The server side now communicates which actions are allowed or not allowed simplifying the logic in the UI to hide the appropriate buttons that are not allowed to be used on a specific client.

Server side this is the same logic used to enforce limits on the API so this ensures that these are kept in sync.